fix compiler warnings in agg building
git-svn-id: http://svn.osgeo.org/mapserver/trunk@11432 7532c77e-422f-0410-93f4-f0b67bdd69e2

+++ renderers/agg/src/agg_svg_parser.cpp
@@ -243,7 +243,7 @@ namespace svg
                 sprintf(msg,
                     "%s at line %d\n",
                     XML_ErrorString(XML_GetErrorCode(p)),
-                    XML_GetCurrentLineNumber(p));
+                    (int)XML_GetCurrentLineNumber(p));
                 throw exception(msg);
             }
         }
